No warnings!!!
changed `linenr_T` to be `int`. This makes more conseptual sense, as I
sincerely doubt any file will be longer than 2^31-1 lines long. Also,
the macro `MAXLNUM` was already defined to be a 31-bit value.

+++ src/nvim/globals.h
@@ -802,7 +802,7 @@ EXTERN int State INIT(= NORMAL);        /* This is the current state of the
                                          * command interpreter. */
 
 EXTERN int finish_op INIT(= FALSE);     /* TRUE while an operator is pending */
-EXTERN long opcount INIT(= 0);          /* count for pending operator */
+EXTERN int opcount INIT(= 0);          /* count for pending operator */
 
 /*
  * ex mode (Q) state
